Update scraper.py
The issue with the provided code is that it doesn't handle the case when the user inputs an invalid option. It will simply continue the loop without any feedback. To fix this, you can add an else block after the last elif statement to handle invalid input and provide feedback to the user.

Another issue is that the code uses quit() to exit the program. It is recommended to use sys.exit() instead to ensure a clean exit. You will need to import the sys module at the beginning of the script.